# openapi.yaml
openapi: 3.0.3
info:
  title: Taskryse API
  version: 1.0.0
  description: >
    Backend contract for Taskryse. Money in minor units (kobo/cents).
    Core rule: plans grant ELIGIBILITY only; earnings come from APPROVED work
    against FUNDED sponsor escrow; promo credits are non-withdrawable.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
paths:
...
  /submissions/{id}/approve: { post: { summary: Approve -> move pending to withdrawable IF escrow funded, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: approved } } } }
  /submissions/{id}/revision: { post: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: revision_requested } } } }
  /submissions/{id}/reject: { post: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: rejected } } } }
  /submissions/{id}/flag: { post: { summary: Flag fraud -> hold + admin risk, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: flagged } } } }
  /submissions/{id}/appeal: { post: { summary: Appeal -> independent second reviewer, parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '201': { description: appeal opened } } } }
...
  /withdrawals:
    post:
      summary: Withdraw. Requires PIN + OTP. Deduct withdrawable, apply fee. States processing->completed|failed_reversed.
      responses: { '201': { description: created }, '402': { description: insufficient withdrawable }, '403': { description: pin/otp failed } }
  /withdrawals/{id}: { get: { parameters: [{name: id, in: path, required: true, schema: {type: string}}], responses: { '200': { description: status } } } }
...